fix geographic models
There was an error in the geographic methods using the improved pscore (ps_spec2). The pscores for the pscore model with state variables used the ps_fit_2 model instead of ps_fit_2.st to get the link predictions. This led to downstream errors in restructuring the data (basically it was restructed the same as w/o state variables), but then the restructured data was used in a TE model with state variables.